PR feedback: Relocate sanitizeTrailingNbspOnPaste from utils.ts to rich_text_editor.ts
