+++ apps/design/frontend/src/rich_text_editor.tsx
@@ -31,7 +31,6 @@ import React, { useState } from 'react';
 import { Buffer } from 'node:buffer';
 import { ImageInputButton } from './image_input';
 import { NormalizeParams } from './image_normalization';
-import { sanitizeTrailingNbspOnPaste } from './utils';
 
 const ControlGroup = styled.div`
   display: flex;
@@ -320,6 +319,34 @@ function unwrapSingleCellTablesOnPaste(slice: Slice): Slice {
     );
   }
   return slice;
+}
+
+// Find the last text node in a given root node
+function findLastTextNode(root: Node): Text | null {
+  const walker = document.createTreeWalker(root, NodeFilter.SHOW_TEXT);
+  let last: Text | null = null;
+  while (walker.nextNode()) last = walker.currentNode as Text;
+  return last;
+}
+
+// Remove trailing NBSPs and trailing whitespace, if there is at least one nbsp (unicode U+00A0)
+function stripTrailingNbsp(text: string): string {
+  return text.replace(/[\u00A0\s]+$/, '');
+}
+
+// HTML_BLOCKS includes the HTML elements that most commonly have unintended trailing
+// non-breaking spaces when pasting content from external sources
+const HTML_BLOCKS = ['p', 'li', 'td', 'th'] as const;
+export function sanitizeTrailingNbspOnPaste(html: string): string {
+  const doc = new DOMParser().parseFromString(html, 'text/html');
+  doc.body.querySelectorAll(HTML_BLOCKS.join(',')).forEach((block) => {
+    const lastText = findLastTextNode(block);
+    if (lastText?.nodeValue) {
+      lastText.nodeValue = stripTrailingNbsp(lastText.nodeValue);
+    }
+  });
+
+  return doc.body.innerHTML;
 }
 
 interface RichTextEditorProps {
